# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FitBid API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success or /api/login on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  # ─── USERS ────────────────────────────────────────────────────────────────
...
  /posts/{postId}/bids:
    get:
      tags: [bids]
      operationId: listBids
      summary: Get bids for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of bids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BidWithProfessional'
    post:
      tags: [bids]
      operationId: createBid
      summary: Submit a bid on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BidInput'
      responses:
        '201':
          description: Bid submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bid'
...
  /bids/{bidId}:
    get:
      tags: [bids]
      operationId: getBid
      summary: Get bid details
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Bid detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidWithProfessional'
    patch:
      tags: [bids]
      operationId: updateBidStatus
      summary: Accept or reject a bid
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BidStatusUpdate'
      responses:
        '200':
          description: Updated bid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bid'

  # ─── ENGAGEMENTS ──────────────────────────────────────────────────────────
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ProfessionalProfile:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
        name:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
        specialty:
          type: [string, 'null']
        bio:
          type: [string, 'null']
        yearsExperience:
          type: [integer, 'null']
        certifications:
          type: array
          items:
            type: string
        verificationStatus:
          type: string
          enum: [pending, approved, rejected]
        avgRating:
          type: [number, 'null']
        totalEngagements:
          type: integer
        hourlyRate:
          type: [number, 'null']
        availabilityStatus:
          type: string
          enum: [open, booked]
        role:
          type: [string, 'null']
          enum: [trainer, nutritionist, null]
...
    Bid:
      type: object
      required: [id, postId, professionalId, status, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        professionalId:
          type: string
        proposalText:
          type: [string, 'null']
        price:
          type: [number, 'null']
        estimatedWeeks:
          type: [integer, 'null']
        introOffer:
          type: [string, 'null']
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
    BidWithProfessional:
      type: object
      required: [id, postId, professionalId, status, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        professionalId:
          type: string
        proposalText:
          type: [string, 'null']
        price:
          type: [number, 'null']
        estimatedWeeks:
          type: [integer, 'null']
        introOffer:
          type: [string, 'null']
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
        professional:
          $ref: '#/components/schemas/ProfessionalProfile'
...
    BidInput:
      type: object
      required: [proposalText, price]
      properties:
        proposalText:
          type: string
          minLength: 1
        price:
          type: number
        estimatedWeeks:
          type: integer
        introOffer:
          type: string
    BidStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected]

    # ─── Engagements ───────────────────────────────────────────────────────
...