# lib/api-zod/src/generated/types/bidStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitBid API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BidStatusUpdateStatus } from './bidStatusUpdateStatus';

export interface BidStatusUpdate {
  status: BidStatusUpdateStatus;
}